+++ packages/analytics-js-plugins/__tests__/utilities/queue.test.ts
@@ -1,25 +1,8 @@
-import type { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
-import type { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
-import {
-  getDeliveryPayload,
-  validateEventPayloadSize,
-} from '@rudderstack/analytics-js-plugins/utilities/eventsDelivery';
-import * as utilConstants from '@rudderstack/analytics-js-plugins/utilities/constants';
-
-class MockLogger implements ILogger {
-  warn = jest.fn();
-  log = jest.fn();
-  error = jest.fn();
-  info = jest.fn();
-  debug = jest.fn();
-  minLogLevel = 0;
-  scope = 'test scope';
-  setMinLogLevel = jest.fn();
-  setScope = jest.fn();
-  logProvider = console;
-}
-
-const mockLogger = new MockLogger();
+import type { RudderContext, RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
+import { defaultLogger } from '@rudderstack/analytics-js-common/__mocks__/Logger';
+import type { ApiObject } from '@rudderstack/analytics-js-common/types/ApiObject';
+import { getDeliveryPayload, validateEventPayloadSize } from '../../src/utilities/eventsDelivery';
+import * as utilConstants from '../../src/utilities/constants';
 
 describe('Queue Plugins Utilities', () => {
   describe('getDeliveryPayload', () => {
@@ -89,7 +72,7 @@ describe('Queue Plugins Utilities', () => {
         properties: {
           test: 'test',
         },
-      };
+      } as unknown as RudderEvent;
 
       expect(getDeliveryPayload(event)).toEqual(
         '{"channel":"test","type":"track","anonymousId":"test","context":{"traits":{"trait_1":"trait_1","trait_2":"trait_2"},"sessionId":1,"sessionStart":true,"consentManagement":{"deniedConsentIds":["1","2","3"]},"ua-ch":{"test":"test"},"app":{"name":"test","version":"test","namespace":"test"},"library":{"name":"test","version":"test"},"userAgent":"test","os":{"name":"test","version":"test"},"locale":"test","screen":{"width":1,"height":1,"density":1,"innerWidth":1,"innerHeight":1},"campaign":{"source":"test","medium":"test","name":"test","term":"test","content":"test"}},"originalTimestamp":"test","integrations":{"All":true},"messageId":"test","previousId":"test","sentAt":"test","category":"test","traits":{"trait_1":"trait_11","trait_2":"trait_12"},"groupId":"test","event":"test","userId":"test","properties":{"test":"test"}}',
@@ -152,7 +135,7 @@ describe('Queue Plugins Utilities', () => {
           test: 'test',
           test2: null,
         },
-      };
+      } as unknown as RudderEvent;
 
       expect(getDeliveryPayload(event)).toEqual(
         '{"channel":"test","type":"track","anonymousId":"test","context":{"traits":{"trait_1":"trait_1","trait_2":"trait_2"},"sessionId":1,"sessionStart":true,"ua-ch":{"test":"test"},"app":{"name":"test","version":"test","namespace":"test"},"library":{"name":"test","version":"test"},"userAgent":"test","os":{"name":"test","version":"test"},"locale":"test","screen":{"width":1,"height":1,"density":1,"innerWidth":1,"innerHeight":1}},"originalTimestamp":"test","integrations":{"All":true},"messageId":"test","previousId":"test","sentAt":"test","category":"test","groupId":"test","event":"test","userId":"test","properties":{"test":"test"}}',
@@ -225,12 +208,12 @@ describe('Queue Plugins Utilities', () => {
         properties: {
           test: 'test',
         },
-      } as RudderEvent;
+      } as unknown as RudderEvent;
 
       event.traits = event.context.traits;
-      event.context.traits.newTraits = event.traits;
+      ((event.context as RudderContext).traits as ApiObject).newTraits = event.traits;
 
-      expect(getDeliveryPayload(event, mockLogger)).toContain('[Circular Reference]');
+      expect(getDeliveryPayload(event, defaultLogger)).toContain('[Circular Reference]');
     });
 
     it('should return null if the payload cannot be stringified', () => {
@@ -238,11 +221,12 @@ describe('Queue Plugins Utilities', () => {
         channel: 'test',
         type: 'track',
         properties: {
+          // eslint-disable-next-line compat/compat
           someBigInt: BigInt(9007199254740991),
         },
       } as unknown as RudderEvent;
 
-      expect(getDeliveryPayload(event, mockLogger)).toBeNull();
+      expect(getDeliveryPayload(event, defaultLogger)).toBeNull();
     });
   });
 
@@ -250,11 +234,17 @@ describe('Queue Plugins Utilities', () => {
     const originalMaxEventPayloadSize = utilConstants.EVENT_PAYLOAD_SIZE_BYTES_LIMIT;
 
     beforeEach(() => {
-      utilConstants.EVENT_PAYLOAD_SIZE_BYTES_LIMIT = 50;
+      Object.defineProperty(utilConstants, 'EVENT_PAYLOAD_SIZE_BYTES_LIMIT', {
+        value: 50,
+        writable: true,
+      });
     });
 
     afterEach(() => {
-      utilConstants.EVENT_PAYLOAD_SIZE_BYTES_LIMIT = originalMaxEventPayloadSize;
+      Object.defineProperty(utilConstants, 'EVENT_PAYLOAD_SIZE_BYTES_LIMIT', {
+        value: originalMaxEventPayloadSize,
+        writable: false,
+      });
     });
 
     it('should log a warning if the payload size is greater than the max limit', () => {
@@ -269,10 +259,10 @@ describe('Queue Plugins Utilities', () => {
         properties: {
           test: 'test',
         },
-      };
-      validateEventPayloadSize(event, mockLogger);
+      } as unknown as RudderEvent;
+      validateEventPayloadSize(event, defaultLogger);
 
-      expect(mockLogger.warn).toHaveBeenCalledWith(
+      expect(defaultLogger.warn).toHaveBeenCalledWith(
         'QueueUtilities:: The size of the event payload (129 bytes) exceeds the maximum limit of 50 bytes. Events with large payloads may be dropped in the future. Please review your instrumentation to ensure that event payloads are within the size limit.',
       );
     });
@@ -281,11 +271,11 @@ describe('Queue Plugins Utilities', () => {
       const event = {
         channel: 'test',
         type: 'track',
-      };
+      } as unknown as RudderEvent;
 
-      validateEventPayloadSize(event, mockLogger);
+      validateEventPayloadSize(event, defaultLogger);
 
-      expect(mockLogger.warn).not.toHaveBeenCalled();
+      expect(defaultLogger.warn).not.toHaveBeenCalled();
     });
 
     it('should not log a warning if the payload size is equal to the max limit', () => {
@@ -295,11 +285,11 @@ describe('Queue Plugins Utilities', () => {
         type: 'track',
         ab: 'd',
         g: 'j',
-      };
+      } as unknown as RudderEvent;
 
-      validateEventPayloadSize(event, mockLogger);
+      validateEventPayloadSize(event, defaultLogger);
 
-      expect(mockLogger.warn).not.toHaveBeenCalled();
+      expect(defaultLogger.warn).not.toHaveBeenCalled();
     });
 
     it('should log a warning if the payload size could not be calculated', () => {
@@ -313,13 +303,14 @@ describe('Queue Plugins Utilities', () => {
         userId: 'test',
         properties: {
           test: 'test',
+          // eslint-disable-next-line compat/compat
           test1: BigInt(9007199254740991),
         },
       } as unknown as RudderEvent;
 
-      validateEventPayloadSize(event, mockLogger);
+      validateEventPayloadSize(event, defaultLogger);
 
-      expect(mockLogger.warn).toHaveBeenCalledWith(
+      expect(defaultLogger.warn).toHaveBeenCalledWith(
         'QueueUtilities:: Failed to validate event payload size. Please make sure that the event payload is within the size limit and is a valid JSON object.',
       );
     });