+++ packages/analytics-js-plugins/__tests__/xhrQueue/utilities.test.ts
@@ -1,8 +1,8 @@
 import type { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
 import type { ResponseDetails } from '@rudderstack/analytics-js-common/types/HttpClient';
-import type { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
-import { state, resetState } from '@rudderstack/analytics-js/state';
 import { getCurrentTimeFormatted } from '@rudderstack/analytics-js-common/utilities/timestamp';
+import { defaultLogger } from '@rudderstack/analytics-js-common/__mocks__/Logger';
+import type { ApiObject } from '@rudderstack/analytics-js-common/types/ApiObject';
 import {
   getNormalizedQueueOptions,
   getDeliveryUrl,
@@ -10,18 +10,14 @@ import {
   logErrorOnFailure,
   getRequestInfo,
   getBatchDeliveryPayload,
-} from '@rudderstack/analytics-js-plugins/xhrQueue/utilities';
+} from '../../src/xhrQueue/utilities';
+import { resetState, state } from '../../__mocks__/state';
 
 jest.mock('@rudderstack/analytics-js-common/utilities/timestamp', () => ({
   getCurrentTimeFormatted: () => '2021-01-01T00:00:00.000Z',
 }));
 
 describe('xhrQueue Plugin Utilities', () => {
-  const mockLogger = {
-    error: jest.fn(),
-    warn: jest.fn(),
-  } as unknown as ILogger;
-
   describe('getNormalizedQueueOptions', () => {
     it('should return default queue options if input queue options is empty object', () => {
       const queueOptions = getNormalizedQueueOptions({});
@@ -36,6 +32,7 @@ describe('xhrQueue Plugin Utilities', () => {
     });
 
     it('should return default queue options if input queue options is null', () => {
+      // @ts-expect-error Testing for null
       const queueOptions = getNormalizedQueueOptions(null);
 
       expect(queueOptions).toEqual({
@@ -48,6 +45,7 @@ describe('xhrQueue Plugin Utilities', () => {
     });
 
     it('should return default queue options if input queue options is undefined', () => {
+      // @ts-expect-error Testing for undefined
       const queueOptions = getNormalizedQueueOptions(undefined);
 
       expect(queueOptions).toEqual({
@@ -122,9 +120,9 @@ describe('xhrQueue Plugin Utilities', () => {
         response: {},
       } as ResponseDetails;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', false, 1, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', false, 1, 10, defaultLogger);
 
-      expect(mockLogger.error).not.toBeCalled();
+      expect(defaultLogger.error).not.toHaveBeenCalled();
     });
 
     it('should log an error for delivery failure', () => {
@@ -132,9 +130,9 @@ describe('xhrQueue Plugin Utilities', () => {
         error: {},
       } as ResponseDetails;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', false, 1, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', false, 1, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. The event(s) will be dropped.',
       );
     });
@@ -147,54 +145,59 @@ describe('xhrQueue Plugin Utilities', () => {
         },
       } as ResponseDetails;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. It/they will be retried. Retry attempt 1 of 10.',
       );
 
       // Retryable error but it's the first attempt
-      details.xhr.status = 429;
+      // @ts-expect-error Needed to set the status for testing
+      (details.xhr as XMLHttpRequest).status = 429;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', true, 0, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', true, 0, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. It/they will be retried.',
       );
 
       // 500 error
-      details.xhr.status = 500;
+      // @ts-expect-error Needed to set the status for testing
+      (details.xhr as XMLHttpRequest).status = 500;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. It/they will be retried. Retry attempt 1 of 10.',
       );
 
       // 5xx error
-      details.xhr.status = 501;
+      // @ts-expect-error Needed to set the status for testing
+      (details.xhr as XMLHttpRequest).status = 501;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. It/they will be retried. Retry attempt 1 of 10.',
       );
 
       // 600 error
-      details.xhr.status = 600;
+      // @ts-expect-error Needed to set the status for testing
+      (details.xhr as XMLHttpRequest).status = 600;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', true, 1, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. The event(s) will be dropped.',
       );
 
       // Retryable error but exhausted all tries
-      details.xhr.status = 520;
+      // @ts-expect-error Needed to set the status for testing
+      (details.xhr as XMLHttpRequest).status = 520;
 
-      logErrorOnFailure(details, 'https://test.com/v1/page', false, 10, 10, mockLogger);
+      logErrorOnFailure(details, 'https://test.com/v1/page', false, 10, 10, defaultLogger);
 
-      expect(mockLogger.error).toBeCalledWith(
+      expect(defaultLogger.error).toHaveBeenCalledWith(
         'XhrQueuePlugin:: Failed to deliver event(s) to https://test.com/v1/page. Retries exhausted (10). The event(s) will be dropped.',
       );
     });
@@ -219,7 +222,7 @@ describe('xhrQueue Plugin Utilities', () => {
         },
       };
 
-      const requestInfo = getRequestInfo(queueItemData, state, mockLogger);
+      const requestInfo = getRequestInfo(queueItemData, state, defaultLogger);
 
       expect(requestInfo).toEqual({
         url: 'https://test.com/v1/track',
@@ -260,7 +263,7 @@ describe('xhrQueue Plugin Utilities', () => {
 
       state.lifecycle.activeDataplaneUrl.value = 'https://test.dataplaneurl.com/';
 
-      const requestInfo = getRequestInfo(queueItemData, state, mockLogger);
+      const requestInfo = getRequestInfo(queueItemData, state, defaultLogger);
 
       expect(requestInfo).toEqual({
         url: 'https://test.dataplaneurl.com/v1/batch',
@@ -294,7 +297,7 @@ describe('xhrQueue Plugin Utilities', () => {
         } as unknown as RudderEvent,
       ];
 
-      expect(getBatchDeliveryPayload(events, currentTime, mockLogger)).toBe(
+      expect(getBatchDeliveryPayload(events, currentTime, defaultLogger)).toBe(
         '{"batch":[{"channel":"test","type":"track","anonymousId":"test","properties":{"test":"test"}},{"channel":"test","type":"track","anonymousId":"test","properties":{"test1":"test1"}}],"sentAt":"2021-01-01T00:00:00.000Z"}',
       );
     });
@@ -324,13 +327,13 @@ describe('xhrQueue Plugin Utilities', () => {
           },
         } as unknown as RudderEvent,
       ];
-      expect(getBatchDeliveryPayload(events, currentTime, mockLogger)).toBe(
+      expect(getBatchDeliveryPayload(events, currentTime, defaultLogger)).toBe(
         '{"batch":[{"channel":"test","type":"track","anonymousId":"test","properties":{"test":"test"}},{"channel":"test","type":"track","anonymousId":"test","properties":{"test1":"test1","test3":{}}}],"sentAt":"2021-01-01T00:00:00.000Z"}',
       );
     });
 
     it('should return string with circular dependencies replaced with static string', () => {
-      const events = [
+      const events: RudderEvent[] = [
         {
           channel: 'test',
           type: 'track',
@@ -354,10 +357,13 @@ describe('xhrQueue Plugin Utilities', () => {
           },
         } as unknown as RudderEvent,
       ];
+      const event2 = events[1] as RudderEvent;
 
-      events[1].properties.test5 = events[1];
+      // Create a circular reference
+      // @ts-expect-error Testing for circular reference
+      (event2.properties as ApiObject).test5 = event2;
 
-      expect(getBatchDeliveryPayload(events, currentTime, mockLogger)).toContain(
+      expect(getBatchDeliveryPayload(events, currentTime, defaultLogger)).toContain(
         '[Circular Reference]',
       );
     });
@@ -382,7 +388,7 @@ describe('xhrQueue Plugin Utilities', () => {
         } as unknown as RudderEvent,
       ];
 
-      expect(getBatchDeliveryPayload(events, currentTime, mockLogger)).toBeNull();
+      expect(getBatchDeliveryPayload(events, currentTime, defaultLogger)).toBeNull();
     });
   });
 });